Full rewrite of the generator
Now Brazil is generated in 25.5s (from 47.6s) and Europe is feasible.

The file format has changed, so this is a breaking change!

The reader part was not yet rewritten, so master is failing in the meanwhile
