+++ Cargo.toml
@@ -9,7 +9,6 @@ edition = "2018"
 [dependencies]
 osmpbf = "0.1.13"
 memmap = "0.7.0"
-rayon = "1.3.0"
 tempfile = "3.1.0"
 crossbeam = "0.7.3"
 petgraph = "0.5.0"
