+++ src/generator/data_types.rs
@@ -3,9 +3,11 @@ mod disk_vec;
 mod graph;
 mod junction;
 mod node;
+mod osm_file;
 
 pub use disk_bit_vec::*;
 pub use disk_vec::*;
 pub use graph::*;
 pub use junction::*;
 pub use node::*;
+pub use osm_file::*;
