+++ src/generator/data_types/osm_file.rs
