+++ src/generator/parser/graph.rs
@@ -3,132 +3,120 @@
 
 use crate::generator::data_types::*;
 use crossbeam;
-use osmpbf::{BlobDecode, MmapBlob, Way};
 
 /// Build the roadmap graph
-pub fn parse_blobs<'a>(
-    blobs: &[MmapBlob],
+pub fn parse_file<'a>(
+    file: &'a OSMClassifiedFile<'a>,
     nodes: &'a Nodes,
     junctions: &'a Junctions,
     num_threads: usize,
-) -> Graph<'a> {
+) -> Graph {
     if num_threads == 1 {
-        parse_blobs_sequential(blobs, nodes, junctions)
+        parse_file_sequential(file, nodes, junctions)
     } else {
-        parse_blobs_parallel(blobs, nodes, junctions, num_threads)
+        parse_file_parallel(file, nodes, junctions, num_threads)
     }
 }
 
 #[derive(Copy, Clone)]
 struct Arc {
-    from: OSMNodeId,
-    to: OSMNodeId,
+    from: NodeIndex,
+    to: NodeIndex,
     road_level: u8,
     distance: u32,
 }
 
 /// Parse the raw ways from a given compressed blob
-fn parse_blob<'a>(blob: &MmapBlob, nodes: &'a Nodes, junctions: &'a Junctions) -> Vec<Arc> {
-    let mut arcs = Vec::new();
-    match blob.decode().unwrap() {
-        BlobDecode::OsmData(block) => {
-            for group in block.groups() {
-                for way in group.ways() {
-                    parse_way(way, nodes, junctions, &mut arcs);
-                }
-            }
-        }
-        _ => {}
-    }
-    arcs
-}
-
 /// Handle each way that is a road, adding arcs into the graph.
 /// First, the way will be split into segments. A segment is a sequence of nodes,
 /// with those at start and end are junction nodes and all the others are non-junctions.
 /// Then, the segment is defined as "blocked" if any of the nodes is a barrier.
 /// Finally, an unblocked segment will push new arcs to the graph. It can push up
 /// to two arcs if the way is both-ways.
-fn parse_way<'a>(way: Way, nodes: &'a Nodes, junctions: &'a Junctions, arcs: &mut Vec<Arc>) {
-    // Parse tags
-    let road_level = match super::parse_road_level(&way) {
-        None => return,
-        Some(x) => x,
-    };
-    let direction = super::parse_oneway(&way);
+fn parse_ways<'a>(ways: &WaysBlob, nodes: &'a Nodes, junctions: &'a Junctions) -> Vec<Arc> {
+    let mut arcs = Vec::new();
+    ways.for_each(|way| {
+        // Parse tags
+        let road_level = match super::parse_road_level(&way) {
+            None => return,
+            Some(x) => x,
+        };
+        let direction = super::parse_oneway(&way);
 
-    let mut it = way.refs();
+        let mut it = way.refs();
 
-    // Handle first node
-    let node_id = it.next().unwrap();
-    let node = nodes.node(node_id).unwrap();
-    let mut seg_start = node;
-    let mut prev_node = node;
-    let mut distance: f64 = 0.;
-    let mut blocked = node.barrier;
-
-    // Handle the other nodes
-    for node_id in it {
+        // Handle first node
+        let node_id = it.next().unwrap();
         let node = nodes.node(node_id).unwrap();
-        distance += prev_node.point.haversine_distance(&node.point);
-        prev_node = node;
-        blocked |= node.barrier;
+        let mut seg_start = node;
+        let mut prev_node = node;
+        let mut distance: f64 = 0.;
+        let mut blocked = node.barrier;
 
-        if junctions.get(node.offset) == NodeType::Junction {
-            if !blocked {
-                // Commit segment
-                if direction.direct {
-                    arcs.push(Arc {
-                        from: seg_start.id,
-                        to: node_id,
-                        road_level,
-                        distance: distance.round() as u32,
-                    });
+        // Handle the other nodes
+        for node_id in it {
+            let node = nodes.node(node_id).unwrap();
+            distance += prev_node.point.haversine_distance(&node.point);
+            prev_node = node;
+            blocked |= node.barrier;
+
+            if junctions.is_junction(node.id) {
+                if !blocked {
+                    // Commit segment
+                    if direction.direct {
+                        arcs.push(Arc {
+                            from: NodeIndex::new(seg_start.offset),
+                            to: NodeIndex::new(node.offset),
+                            road_level,
+                            distance: distance.round() as u32,
+                        });
+                    }
+                    if direction.reverse {
+                        arcs.push(Arc {
+                            from: NodeIndex::new(node.offset),
+                            to: NodeIndex::new(seg_start.offset),
+                            road_level,
+                            distance: distance.round() as u32,
+                        });
+                    }
                 }
-                if direction.reverse {
-                    arcs.push(Arc {
-                        from: node_id,
-                        to: seg_start.id,
-                        road_level,
-                        distance: distance.round() as u32,
-                    });
-                }
+                seg_start = node;
+                distance = 0.;
+                blocked = node.barrier;
             }
-            seg_start = node;
-            distance = 0.;
-            blocked = node.barrier;
         }
-    }
 
-    // By definition, the last node is a junction, so the last segment will be commited
-    assert_eq!(distance, 0.);
+        // By definition, the last node is a junction, so the last segment will be commited
+        assert_eq!(distance, 0.);
+    });
+    arcs
 }
 
-fn parse_blobs_sequential<'a>(
-    blobs: &[MmapBlob],
+fn parse_file_sequential<'a>(
+    file: &'a OSMClassifiedFile<'a>,
     nodes: &'a Nodes,
     junctions: &'a Junctions,
-) -> Graph<'a> {
+) -> Graph {
     let mut graph = Graph::new(nodes);
-    for blob in blobs {
-        for arc in parse_blob(blob, nodes, junctions) {
+    for ways in &file.ways_blobs {
+        for arc in parse_ways(ways, nodes, junctions) {
             graph.push_arc(arc.from, arc.to, arc.road_level, arc.distance);
         }
     }
     graph
 }
 
-fn parse_blobs_parallel<'a>(
-    blobs: &[MmapBlob],
+fn parse_file_parallel<'a>(
+    file: &'a OSMClassifiedFile<'a>,
     nodes: &'a Nodes,
     junctions: &'a Junctions,
     num_threads: usize,
-) -> Graph<'a> {
+) -> Graph {
     crossbeam::scope(|scope| {
         // Create a work queue that will be filled once by this thread and will be
         // consumed by the worker ones.
-        let (task_sender, task_receiver) = crossbeam::bounded(blobs.len());
-        for task in blobs {
+        let (task_sender, task_receiver) = crossbeam::bounded(file.ways_blobs.len());
+        for task in &file.ways_blobs {
             task_sender.send(task).unwrap();
         }
         drop(task_sender);
@@ -142,14 +130,15 @@ fn parse_blobs_parallel<'a>(
             let task_receiver = task_receiver.clone();
             let result_sender = result_sender.clone();
             scope.spawn(move |_| {
-                for blob in task_receiver {
+                for ways in task_receiver {
                     result_sender
-                        .send(parse_blob(blob, nodes, junctions))
+                        .send(parse_ways(ways, nodes, junctions))
                         .unwrap();
                 }
             });
         }
         drop(result_sender);
+
         // Consume the results and push then in order to the storage builder
         let mut graph = Graph::new(nodes);
         for arcs in result_receiver {
