--- src/generator2/data_types.rs
@@ -1,13 +1,0 @@
-mod disk_bit_vec;
-mod disk_vec;
-mod graph;
-mod junction;
-mod node;
-mod osm_file;
-
-pub use disk_bit_vec::*;
-pub use disk_vec::*;
-pub use graph::*;
-pub use junction::*;
-pub use node::*;
-pub use osm_file::*;
